third_party/WebKit/LayoutTests/platform/linux/virtual/incremental-shadow-dom/external/wpt/css/css-scoping/slotted-invalidation-expected.txt

This is a testharness.js-based test.
FAIL CSS Test: Style invalidation for ::slotted() assert_equals: expected "rgb(255, 0, 0)" but got "rgba(0, 0, 0, 0)"
Harness: the test ran to completion.
